Split Wagtail models into base classes and page models, ahead of adding custom image models
